requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
crewai>=0.1.0
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.25.0
beautifulsoup4>=4.10.0
